renamed omdbApi.ts to movieService.ts in frontend and added constants
