+++ frontend/src/constants/appConstants.ts
@@ -6,6 +6,10 @@ export const FE_LOGIN_FAILED = "Login failed. Please check your credentials.";
 export const FE_REGISTER_FAILED = "Registration failed. Please try again.";
 export const FE_FETCH_USER_FAILED = "Failed to fetch user data.";
 
+// App Details Constants
+export const APP_NAME = "Movie Recommendation App";
+export const APP_DESCRIPTION = "Discover and manage your favorite movies with personalized recommendations.";
+
 // Local Storage Keys
 export const LS_AUTH_TOKEN = "token";
 export const LS_USER_DATA = "user";
@@ -18,9 +22,10 @@ export const API_AUTH_LOGOUT = "/auth/logout";
 
 // Frontend Route Paths
 export const PATH_LOGIN = "/login";
+export const PATH_HOME = "/";
 export const PATH_REGISTER = "/register";
 export const PATH_DASHBOARD = "/dashboard";
-export const PATH_HOME = "/";
+export const PATH_MOVIE_DETAIL = "/movie/:imdbID";
 
 // Frontend Form Validation Messages
 export const VALIDATION_EMAIL_REQUIRED = "Email is required";
@@ -32,7 +37,7 @@ export const VALIDATION_PASSWORD_MIN_LENGTH = "Password must be at least 6 chara
 export const VALIDATION_CONFIRM_PASSWORD_REQUIRED = "Please confirm your password";
 export const VALIDATION_PASSWORDS_DO_NOT_MATCH = "Passwords do not match";
 
-// OMDB API
-export const OMDB_API_BASE_URL = "http://www.omdbapi.com/";
+// OMDB API Constants
 export const FE_OMDB_FETCH_FAILED = "Failed to fetch movie data from OMDb. Please try again later.";
 export const FE_OMDB_NO_RESULTS = "No movies found matching your search.";
+export const FE_OMDB_DEFAULT_QUERY = "superman";
